backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
absl-py==2.3.1
    # via
    #   keras
    #   tensorboard
    #   tensorflow
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via starlette
astunparse==1.6.3
    # via tensorflow
blinker==1.9.0
    # via flask
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   flask
    #   uvicorn
colorama==0.4.6
    # via click
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dnspython==2.8.0
    # via pymongo
fastapi==0.116.1
    # via -r requirements.in
filelock==3.19.1
    # via torch
flask==3.1.2
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-jwt-extended
    #   flask-pymongo
flask-cors==6.0.1
    # via -r requirements.in
flask-jwt-extended==4.7.1
    # via -r requirements.in
flask-pymongo==3.0.1
    # via -r requirements.in
flatbuffers==25.2.10
    # via tensorflow
fonttools==4.59.2
    # via matplotlib
fsspec==2025.9.0
    # via torch
gast==0.6.0
    # via tensorflow
google-pasta==0.2.0
    # via tensorflow
grpcio==1.74.0
    # via
    #   tensorboard
    #   tensorflow
gunicorn==23.0.0
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
h5py==3.14.0
    # via
    #   keras
    #   tensorflow
idna==3.10
    # via
    #   anyio
    #   requests
imageio==2.37.0
    # via scikit-image
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   flask
    #   torch
joblib==1.5.2
    # via scikit-learn
keras==3.11.3
    # via tensorflow
kiwisolver==1.4.9
    # via matplotlib
lazy-loader==0.4
    # via scikit-image
libclang==18.1.1
    # via tensorflow
markdown==3.9
    # via tensorboard
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
matplotlib==3.10.6
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
ml-dtypes==0.5.3
    # via
    #   keras
    #   tensorflow
mpmath==1.3.0
    # via sympy
namex==0.1.0
    # via keras
networkx==3.5
    # via
    #   scikit-image
    #   torch
numpy==2.3.3
    # via
    #   contourpy
    #   h5py
    #   imageio
    #   keras
    #   matplotlib
    #   ml-dtypes
    #   pandas
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   tensorflow
    #   tifffile
    #   torchvision
opt-einsum==3.4.0
    # via tensorflow
optree==0.17.0
    # via keras
packaging==25.0
    # via
    #   gunicorn
    #   keras
    #   lazy-loader
    #   matplotlib
    #   scikit-image
    #   tensorboard
    #   tensorflow
pandas==2.3.2
    # via -r requirements.in
pillow==11.3.0
    # via
    #   imageio
    #   matplotlib
    #   scikit-image
    #   tensorboard
    #   torchvision
protobuf==6.32.0
    # via
    #   tensorboard
    #   tensorflow
pydantic==2.11.7
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via rich
pyjwt==2.10.1
    # via flask-jwt-extended
pymongo==4.14.1
    # via flask-pymongo
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.1.1
    # via -r requirements.in
pytz==2025.2
    # via pandas
requests==2.32.5
    # via
    #   -r requirements.in
    #   tensorflow
rich==14.1.0
    # via keras
scikit-image==0.25.2
    # via -r requirements.in
scikit-learn==1.7.2
    # via -r requirements.in
scipy==1.16.1
    # via
    #   scikit-image
    #   scikit-learn
six==1.17.0
    # via
    #   astunparse
    #   google-pasta
    #   python-dateutil
    #   tensorflow
sniffio==1.3.1
    # via anyio
starlette==0.47.3
    # via fastapi
sympy==1.14.0
    # via torch
tensorboard==2.20.0
    # via tensorflow
tensorboard-data-server==0.7.2
    # via tensorboard
tensorflow==2.20.0
    # via -r requirements.in
termcolor==3.1.0
    # via tensorflow
threadpoolctl==3.6.0
    # via scikit-learn
tifffile==2025.9.9
    # via scikit-image
torch==2.7.0
    # via
    #   -r requirements.in
    #   torchvision
torchvision==0.22.0
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   optree
    #   pydantic
    #   pydantic-core
    #   starlette
    #   tensorflow
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via -r requirements.in
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
    #   flask-jwt-extended
    #   tensorboard
wheel==0.45.1
    # via astunparse
wrapt==1.17.3
    # via tensorflow

# The following packages are considered to be unsafe in a requirements file:
# setuptools
